backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
apscheduler>=3.10.0
yfinance>=0.2.40
requests>=2.32.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0
pytz>=2024.1
httpx>=0.27.0
pytest>=8.0.0
pydantic-settings>=2.0.0
pandas>=2.0,<4.0
numpy>=1.24,<3.0
boto3>=1.34.0
